lib: move GetMyPath from config to proc
