+++ lib/config.cpp
@@ -890,72 +890,6 @@ void Util::Config::addStandardPushCapabilities(JSON::Value &cap){
 
 }
 
-/// Gets directory the current executable is stored in.
-std::string Util::getMyPath(){
-  char mypath[500];
-#ifdef __CYGWIN__
-  GetModuleFileName(0, mypath, 500);
-#else
-#ifdef __APPLE__
-  memset(mypath, 0, 500);
-  unsigned int refSize = 500;
-  _NSGetExecutablePath(mypath, &refSize);
-#else
-  int ret = readlink("/proc/self/exe", mypath, 500);
-  if (ret != -1){
-    mypath[ret] = 0;
-  }else{
-    mypath[0] = 0;
-  }
-#endif
-#endif
-  std::string tPath = mypath;
-  size_t slash = tPath.rfind('/');
-  if (slash == std::string::npos){
-    slash = tPath.rfind('\\');
-    if (slash == std::string::npos){return "";}
-  }
-  tPath.resize(slash + 1);
-  return tPath;
-}
-
-/// Gets all executables in getMyPath that start with "Mist".
-void Util::getMyExec(std::deque<std::string> &execs){
-  std::string path = Util::getMyPath();
-#ifdef __CYGWIN__
-  path += "\\Mist*";
-  WIN32_FIND_DATA FindFileData;
-  HANDLE hdl = FindFirstFile(path.c_str(), &FindFileData);
-  while (hdl != INVALID_HANDLE_VALUE){
-    if (!(FindFileData.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY)){
-      execs.push_back(FindFileData.cFileName);
-    }
-    if (!FindNextFile(hdl, &FindFileData)){
-      FindClose(hdl);
-      hdl = INVALID_HANDLE_VALUE;
-    }
-  }
-#else
-  DIR *d = opendir(path.c_str());
-  if (!d){return;}
-  struct dirent *dp;
-  do{
-    errno = 0;
-    if ((dp = readdir(d))){
-      if (dp->d_type != DT_DIR && strncmp(dp->d_name, "Mist", 4) == 0){
-        if (dp->d_type != DT_REG) {
-          struct stat st = {};
-          stat(dp->d_name, &st);
-          if (!S_ISREG(st.st_mode))
-            continue;
-        }
-        execs.push_back(dp->d_name);}
-    }
-  }while (dp != NULL);
-  closedir(d);
-#endif
-}
-
 /// Sets the current process' running user
 void Util::setUser(std::string username){
   if (username != "root"){
